+++ R/estimation.R
@@ -1,11 +1,11 @@
 #' Estimation and estimation accuracy for a pqrBayes object
 #'
-#' Calculate estimated regression coefficients with estimation accuracy from linear and quantile VC models, respectively.
+#' Calculate estimated regression coefficients with estimation accuracy from linear, binary LASSO, group LASSO and quantile VC models, respectively.
 #'
 #' @param object an object of class `pqrBayes'.
-#' @param coefficient the vector of quantile regression coefficients under a linear model or the matrix of true varying coefficients evaluated on the grid points under a varying coefficient model.
-#' @param u.grid the vector of grid points under a varying coefficient model. When fitting a linear (quantile) regression model or group LASSO, u.grid = NULL.
-#' @param model the model to be fitted. Users can choose "linear" for a linear model, "VC" for a varying coefficient model or "group for group LASSO.
+#' @param coefficient the vector of quantile regression coefficients under a linear model (i.e., LASSO), binary LASSO and group LASSO or the matrix of true varying coefficients evaluated on the grid points under a varying coefficient model.
+#' @param u.grid the vector of grid points under a varying coefficient model. When fitting a linear regression model (i.e., LASSO), binary LASSO or group LASSO, u.grid = NULL.
+#' @param model the model to be fitted. Users can choose "linear" for a linear model (i.e., LASSO), "binary" for binary LASSO, "group" for group LASSO or "VC" for a varying coefficient model.
 #' @usage estimation.pqrBayes(object,coefficient,u.grid=NULL,model="linear")
 #' @return  an object of class `pqrBayes.est' is returned, which is a list with components:
 #' \item{error}{mean square error or integrated mean square errors and total integrated mean square error.}
@@ -31,11 +31,14 @@ estimation.pqrBayes = function(object,coefficient,u.grid=NULL,model="linear"){
   else if(model=="linear"){
     pqrBayes.est = estimation_lin(object,coefficient)
   }
+  else if(model=="binary"){
+    pqrBayes.est = estimation_lin(object,coefficient)
+  }
   else if(model=="group"){
     pqrBayes.est = estimation_lin(object,coefficient)
   }
   else{
-    stop("model should be either VC, linear or group")
+    stop("model should be either VC, linear,binary or group")
   }
   class(pqrBayes.est) = "pqrBayes.est"
   return(pqrBayes.est)